Rename routes to component name
