+++ src/pages/tools/NorwegianTariffCreatorPage.tsx
