+++ src/pages/tools/SwitcherAutomationPage.tsx
